fix: dto, response

+++ src/reviews/reviews.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { UpdateProductReviewDto } from 'src/products/dto/updateProductDto';
 import { ReviewsRepository } from './reviews.repository';
 
@@ -9,9 +9,19 @@ export class ReviewsService {
   // 상품 리뷰 수정
   async update(
     id: string,
+    userId: string,
     updateProductReviewDto: UpdateProductReviewDto,
   ): Promise<object> {
-    const review = await this.reviewsRepository.updateProductReview(
+    const review = await this.reviewsRepository.findByReviewId(id);
+
+    if (review.userId !== userId) {
+      throw new HttpException(
+        '다른 사용자의 리뷰를 수정할 권한이 없습니다.',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
+    const newReview = await this.reviewsRepository.updateProductReview(
       id,
       updateProductReviewDto,
     );
@@ -19,18 +29,27 @@ export class ReviewsService {
     return {
       success: true,
       data: {
-        id: review.id,
-        rating: review.rating,
-        title: review.title,
-        content: review.content,
-        updated_at: review.updatedAt,
+        id: newReview.id,
+        rating: newReview.rating,
+        title: newReview.title,
+        content: newReview.content,
+        updated_at: newReview.updatedAt,
       },
       message: '리뷰가 성공적으로 수정되었습니다.',
     };
   }
 
   // 상품 리뷰 삭제
-  async delete(id: string): Promise<object> {
+  async delete(id: string, userId: string): Promise<object> {
+    const review = await this.reviewsRepository.findByReviewId(id);
+
+    if (review.userId !== userId) {
+      throw new HttpException(
+        '다른 사용자의 리뷰를 수정할 권한이 없습니다.',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
     await this.reviewsRepository.deleteProductReview(id);
 
     return {
